+++ src/Products/urban/migration/update_260.py
@@ -1,0 +1,13 @@
+from plone import api
+import logging
+
+
+def add_couple_to_preliminary_notice(context):
+    """
+    """
+    logger = logging.getLogger('urban: add Couple to Preliminary Notice')
+    logger.info("starting upgrade steps")
+    setup_tool = api.portal.get_tool('portal_setup')
+    setup_tool.runImportStepFromProfile('profile-Products.urban:preinstall', 'typeinfo')
+    setup_tool.runImportStepFromProfile('profile-Products.urban:preinstall', 'workflow')
+    logger.info("upgrade step done!")